Env.FD: use var instead of new() for clarity
Change the C.mdb_filehandle_t declaration to use var instead of new() to
clarify that we are not retaining a pointer to anything.

This is related to #28, but in this case no issue, because we are not
assiging a pointer to Go allocated memory to a C value.

Functionally this does not change anything. Go decides if values are
stack allocated or heap allocated based on if they escape, not based on
the syntax.